+++ src/code/_content.js
@@ -1,7 +1,8 @@
 import './compatibility'
-import LandmarksFinder from './landmarksFinder.js'
-import ElementFocuser from './elementFocuser.js'
-import PauseHandler from './pauseHandler.js'
+import LandmarksFinder from './landmarksFinder'
+import ElementFocuser from './elementFocuser'
+import PauseHandler from './pauseHandler'
+import Logger from './logger'
 
 const outOfDateTime = 2000
 const logger = new Logger()
@@ -10,53 +11,6 @@ let observer = null
 const lf = new LandmarksFinder(window, document)
 const ef = new ElementFocuser()
 const ph = new PauseHandler(logger)
-
-
-//
-// Log messages according to user setting
-//
-
-function Logger() {
-	const that = this
-
-	function getDebugInfoOption(callback) {
-		browser.storage.sync.get({
-			debugInfo: false
-		}, function(items) {
-			// We only define the log() function after successfully initing, so
-			// as to trap any errant uses of the logger.
-			handleOptionsChange({
-				debugInfo: {
-					newValue: items.debugInfo
-				}
-			})
-			if (callback) {
-				callback()
-			}
-		})
-	}
-
-	function handleOptionsChange(changes) {
-		if (changes.hasOwnProperty('debugInfo')) {
-			// Ensure the correct line number is reported
-			// https://stackoverflow.com/a/32928812/1485308
-			// https://stackoverflow.com/a/28668819/1485308
-			if (changes.debugInfo.newValue === true) {
-				that.log = console.log.bind(window.console)
-			} else {
-				that.log = function() {}
-			}
-		}
-	}
-
-	// We may wish to log messages right way, but the call to get the user
-	// setting is asynchronous. Therefore, we need to pass our bootstrapping
-	// code as a callback that is run when the option has been fetched.
-	this.init = function(callback) {
-		getDebugInfoOption(callback)
-		browser.storage.onChanged.addListener(handleOptionsChange)
-	}
-}
 
 
 //
