+++ src/code/landmarkName.js
@@ -1,6 +1,6 @@
 // If the landmark has a label, the name is: 'label (role)'
 // otherwise the name is just 'role'
-export default function(landmark) {
+export default function landmarkName(landmark) {
 	if (landmark.label) {
 		return landmark.label + ' (' + processRole(landmark.role) + ')'
 	}
