+++ src/code/sendToActiveTab.js
@@ -1,4 +1,4 @@
-export default function(message, callback) {
+export default function sendToActiveTab(message, callback) {
 	browser.tabs.query({ active: true, currentWindow: true }, function(tabs) {
 		browser.tabs.sendMessage(tabs[0].id, message, callback)
 	})
